feat: init spring security and delete chat

+++ User_Server/src/main/java/com/example/user_server/repository/RoleRepository.java
@@ -1,12 +1,10 @@
 package com.example.user_server.repository;
 
-
 import com.example.user_server.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role,Long> {
-
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByCode(String code);
 }
